backend/requirements.txt
Django==5.0.1
django-ninja==1.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
